cli: Expand package list when bug is sanity-check+ already
While the original intent was to avoid touching bugs once they've
started being worked on, I can't think of a way of doing that without
adding a lot of complexity.  The current behavior of leaving bugs intact
causes a lot of confusion once arches are CC-ed.  Let's expand it
as long as special tokens are there, the bug is sanity-check+ and arch
teams are in CC.

Closes: https://github.com/mgorny/nattka/issues/37
Signed-off-by: Michał Górny <mgorny@gentoo.org>