Revert "package: Expand "*" to empty if it matches CC"
This reverts commit 2e079dd3463f2d5ad8005d42a3e83e094aaa9325.

Given a package list containing `=category/package-${PV}*`, nattka will
remove the `*` and not replace it with a list of architecture keywords
iff the set of architectures to be CC'd matches the set of stable
keywords for `=category/package-${PV}`.

This behavior was added in commit 2e079dd ("package: Expand "*" to empty
if it matches CC") to allow reusing a stabilization bug for an
additional architecture by only adding the new architecture to the CC
list.

This has the unfortunate side-effect of making a nattka-processed
package list unacceptable to nattka itself if the package list needs to
be modified. See https://bugs.gentoo.org/920053#c18 for example.

Revert to the original behavior of always expanding `*` to the
architecture list. If a stabilization needs to happen on an additional
architecture, the additional architecture keyword should be added to the
package list so nattka can sanity-check it.

Signed-off-by: Matt Turner <mattst88@gentoo.org>